[python] extract converter_stmt.cpp from python_converter.cpp
Move 23 statement/assignment methods of python_converter into
src/python-frontend/converter/converter_stmt.cpp: adjust_statement_types,
extract_type_info, create_lhs_expression, handle_assignment_type_adjustments,
handle_array_unpacking, handle_list_literal_unpacking,
get_rhs_with_dict_resolution, infer_type_from_any_annotation,
handle_unpacking_assignment, create_symbol_for_unannotated_assign,
handle_function_call_rhs, handle_string_literal_rhs, is_global_variable,
extract_target_name, preregister_global_variables, get_var_assign,
resolve_variable_type, get_compound_assign, get_conditional_stm,
get_return_statements, get_block, get_static_array, get_delete_statement.

Promote get_statement_type and resolve_ternary_type into
converter/converter_internal.h so the residual get_return_statement
free helper in python_converter.cpp keeps compiling.

Pure code movement — no logic change.
